Moved shield-study-type-specifics to separate class, added corresponding stub class for pioneer
